Split smoke tests for @Stepwise into 3 categories (specs)
